+++ src/dog_marker/api/v1/endpoints/dependecies/__init__.py
@@ -1,6 +1,8 @@
-__all__ = ["get_db", "get_service", "query_coordinate", "query_pagination"]
+__all__ = ["get_config", "get_db", "get_service", "query_coordinate", "query_pagination", "authenticate_app"]
 
-from .get_db import get_db
-from .get_service import get_service
-from .query_coordinate import query_coordinate
-from .query_pagination import query_pagination
+from .auth import authenticate_app
+from .config import get_config
+from .db import get_db
+from .service import get_service
+from .coordinate import query_coordinate
+from .pagination import query_pagination