+++ src/dog_marker/api/v1/endpoints/dependecies/config.py
@@ -1,0 +1,7 @@
+from fastapi import Request
+
+from dog_marker import Config
+
+
+def get_config(request: Request) -> Config:
+    return request.state.config