+++ src/dog_marker/api/v1/endpoints/dependecies/service.py
@@ -3,7 +3,7 @@ from typing import Callable, TypeVar
 from fastapi import Depends
 from sqlalchemy.orm import Session
 
-from .get_db import get_db
+from .db import get_db
 
 T = TypeVar("T")