+++ src/dog_marker/configs.py
@@ -8,6 +8,8 @@ class Config:
     POSTGRES_DB_POOL_SIZE: int = int(os.environ.get("POSTGRES_DB_POOL_SIZE", "20"))
     POSTGRES_DB_MAX_OVERFLOW: int = int(os.environ.get("POSTGRES_DB_MAX_OVERFLOW", "20"))
 
+    APP_TOKEN: str | None = os.environ.get("APP_TOKEN", None)
+
 
 class DevelopConfig(Config):
     pass